+++ front/src/app/[locale]/account/[id]/page.tsx
@@ -4,31 +4,16 @@ import { client } from "@/auth";
 import useSWR from "swr";
 import * as Mantine from "@mantine/core";
 import { useTranslations } from "next-intl";
-import { useState } from "react";
-
-const noticeStatesDummy = {
-  app: {
-    favorite: false,
-    bookmark: false,
-    comment: false,
-    follow: false,
-  },
-  email: {
-    favorite: false,
-    bookmark: false,
-    comment: false,
-    follow: false,
-  },
-};
+import { NoticeTabs } from "@/components/features/users";
 
 const fetcher = (url: string) => client.get(url).then((res) => res.data);
 
 export default function AccountPage() {
   const { data, error } = useSWR("/account", fetcher);
-  console.log(data, error);
   const t_AccountSettings = useTranslations("AccountSettings");
-  const [noticeStates, setNoticeStates] =
-    useState<NoticeStates>(noticeStatesDummy);
+  const t_General = useTranslations("General");
+  // TODO : ローディング画面
+  if (error || data === undefined) return <div>Now Loading</div>;
 
   return (
     <article className="my-8 m-auto w-full px-4">
@@ -42,13 +27,13 @@ export default function AccountPage() {
               {t_AccountSettings("accountName")}
             </dt>
             <dd className="ml-4 md:ml-0 border-b border-slate-300 pb-2">
-              アカウント名
+              {data.account.name}
             </dd>
             <dt className="md:border-b md:border-slate-300 md:pb-2">
               {t_AccountSettings("email")}
             </dt>
             <dd className="flex flex-col ml-4 md:ml-0 border-b border-slate-300 pb-2">
-              *****@example.com
+              {data.account.email ?? data.account.google | data.account.discord}
               <div>
                 <Mantine.Button
                   variant="contained"
@@ -72,143 +57,12 @@ export default function AccountPage() {
           <h3 className="text-xl font-semibold text-center mt-10 mb-2">
             {t_AccountSettings("notificationSettings")}
           </h3>
-          <Tabs noticeStates={noticeStates} setNoticeStates={setNoticeStates} />
+          <NoticeTabs noticeStates={data.notices} />
           <div className="text-center">
-            <Mantine.Button>保存</Mantine.Button>
+            <Mantine.Button>{t_General("save")}</Mantine.Button>
           </div>
         </section>
       </div>
     </article>
-  );
-}
-
-interface NoticeState {
-  favorite: boolean;
-  bookmark: boolean;
-  comment: boolean;
-  follow: boolean;
-}
-
-enum NoticeType {
-  app = "app",
-  email = "email",
-}
-
-interface NoticeStates {
-  app: NoticeState;
-  email: NoticeState;
-}
-
-interface TabPanelProps {
-  children?: React.ReactNode;
-  index: string;
-  value: string;
-  noticeState: NoticeState;
-  setNoticeStates: React.Dispatch<React.SetStateAction<NoticeStates>>;
-}
-
-function Tabs({
-  noticeStates,
-  setNoticeStates,
-}: {
-  noticeStates: NoticeStates;
-  setNoticeStates: React.Dispatch<React.SetStateAction<NoticeStates>>;
-}) {
-  const [value, setValue] = useState<string>(NoticeType.app);
-  const t_AccountSettings = useTranslations("AccountSettings");
-
-  const handleChange = (newValue: string | null) => {
-    if (newValue === value) return;
-
-    if (newValue === null) {
-      setValue(NoticeType.app);
-      return;
-    }
-    setValue(newValue);
-  };
-
-  return (
-    <Mantine.Box>
-      <Mantine.Tabs value={value} onChange={handleChange}>
-        <Mantine.Tabs.List
-          aria-label={t_AccountSettings("notificationSettings")}
-        >
-          <Mantine.Tabs.Tab value={NoticeType.app}>
-            {t_AccountSettings("app")}
-          </Mantine.Tabs.Tab>
-          <Mantine.Tabs.Tab value={NoticeType.email}>
-            {t_AccountSettings("mail")}
-          </Mantine.Tabs.Tab>
-        </Mantine.Tabs.List>
-        <Mantine.Tabs.Panel value={NoticeType.app} className="my-5">
-          <CustomTabPanel
-            value={value}
-            index={NoticeType.app}
-            noticeState={noticeStates.app}
-            setNoticeStates={setNoticeStates}
-          />
-        </Mantine.Tabs.Panel>
-        <Mantine.Tabs.Panel value={NoticeType.email} className="my-5">
-          <CustomTabPanel
-            value={value}
-            index={NoticeType.email}
-            noticeState={noticeStates.email}
-            setNoticeStates={setNoticeStates}
-          />
-        </Mantine.Tabs.Panel>
-      </Mantine.Tabs>
-    </Mantine.Box>
-  );
-}
-
-function CustomTabPanel(props: TabPanelProps) {
-  const { value, noticeState, index, setNoticeStates, ...other } = props;
-  const t_AccountSettings = useTranslations("AccountSettings");
-
-  const handleChange = (key: string, value: boolean) => {
-    const newNoticeState = {
-      ...noticeState,
-      [key]: value,
-    };
-    setNoticeStates((prevState) => ({
-      ...prevState,
-      [index]: newNoticeState,
-    }));
-  };
-
-  return (
-    <div
-      role="tabpanel"
-      hidden={value !== index}
-      id={`tabpanel-${index}`}
-      aria-labelledby={`tab-${index}`}
-      {...other}
-      className="flex flex-col gap-2"
-    >
-      {value === index && (
-        <>
-          <Mantine.Switch
-            label={t_AccountSettings("favorite")}
-            checked={noticeState.favorite}
-            onChange={() => handleChange("favorite", !noticeState.favorite)}
-          />
-          <Mantine.Switch
-            label={t_AccountSettings("bookmark")}
-            checked={noticeState.bookmark}
-            onChange={() => handleChange("bookmark", !noticeState.bookmark)}
-          />
-          <Mantine.Switch
-            label={t_AccountSettings("comment")}
-            checked={noticeState.comment}
-            onChange={() => handleChange("comment", !noticeState.comment)}
-          />
-          <Mantine.Switch
-            label={t_AccountSettings("follow")}
-            checked={noticeState.follow}
-            onChange={() => handleChange("follow", !noticeState.follow)}
-          />
-        </>
-      )}
-    </div>
   );
 }
