+++ front/src/components/features/users/index.ts
@@ -1,4 +1,5 @@
 import UserTabs from "./userTabs";
 import UserEdit from "./edit";
+import NoticeTabs from "./notice";
 
-export { UserTabs, UserEdit };
+export { UserTabs, UserEdit, NoticeTabs };
